Move error trace sanitization from IAST into log collector
